==> Codes/0_1_ls_packages.R <==
# == data loading == #
library(here)
library(downloader)

# == spatial management == #
library(terra)
library(sf)
library(raster)
library(exactextractr)
library(sp)


# == weather == #
library(SPEI)
library(geosphere)


# == data wangling == #
library(tidyverse)
library(data.table)
library(parallel)
library(lubridate)


# ==  visualization == #
library(ggplot2)
library(tmap)


# == ML == #
library(grf)

# === Python === #
library(reticulate)
