==> Codes/1_7_data_exploration.rmd <==
---
title: "Data Exploration"
author: "Shunkei Kakimoto"
output:
  html_document:
    number_sections: yes
    theme: cerulean
    toc_float: yes
    toc: yes
    toc_depth: 3
geometry: margin=1in
---

```{r setup, include=FALSE}

library(knitr)
library(here)

here::i_am("GitControlled/Codes/1_5_data_exploration.rmd")

knitr::opts_chunk$set(
  echo = TRUE,
  cache = FALSE,
  comment = NA,
  message = FALSE,
  warning = FALSE,
  tidy = FALSE,
  cache.lazy = FALSE,
  #--- figure ---#
  dpi = 400,
  fig.width = 7.5,
  fig.height = 5,
  out.width = "750px",
  out.height = "500px"
)

# /*===== Basic Packages  =====*/
# /*---- Data Wrangling ----*/
library(data.table)
library(tidyverse)
library(DescTools)
library(maps)

# /*---- Visualization ----*/
library(RColorBrewer)
library(patchwork)
library(ggplot2)
library(ggthemes)
library(ggpubr)
library(viridis)
library(grid)
library(gridExtra)
library(GGally)

# /*---- Model Summary ----*/
library(stats)
library(modelsummary)
library(gt)
library(flextable)
library(officer)
library(officedown)
```

```{r, echo=FALSE}
custom_theme <- function(x, ...) {
    x %>% gt::opt_row_striping(row_striping = TRUE)
}
options("modelsummary_theme_gt" = custom_theme)

#--- define a function ---#
MinMax <- function(x){
  paste0('[', min(x, na.rm = TRUE) %>% round(digits = 2), ', ', max(x, na.rm = TRUE) %>% round(digits = 2), ']')
} 
```


# Objective
+ check the new regression data (`comp_reg_dt.rmd`)
	* report the summary statistics and visualize the data

## Question

+ `usage` contains *inf*. `usage <= 40` is used to filter out such observations, but is it correct? Why `usage <= 40` is selected although there are some observations whose `usage` is more than *40*?

<br>

<br>

# Data Exploration

## Loading data

```{r}
# /*===== load packages =====*/
source(here("GitControlled/Codes/0_1_ls_packages.R"))

# /*===== load data =====*/
# /*---- newly updated data ----*/
reg_data <- readRDS(here("Shared/Data/WaterAnalysis/comp_reg_dt.rds")) %>%
  .[year %in% 2008:2012 & usage <= 40,]

# /*---- newly updated aggregated data ()----*/
# reg_agg_data <- readRDS(here("/Shared/Data/WaterAnalysis/agg_comp_reg_dt.rds"))

# /*---- old data ----*/
# var_ls_old <- c('precip_in', 'tmin_in', 'tmax_in', 'gdd_in', 'sand_pct', 'silt_pct', 'clay_pct', 'slope', 'kv', 'awc','usage','treat2', 'tr', 'year', 'nrdname')

# reg_data_old <- readRDS(here("/Shared/Data/WaterAnalysis/data_w_LR_TB.rds")) %>%
# 	.[,..var_ls_old]
```

<br>

<br>

## Weather and Soil variables: Site-Level
+ remove `tmmn_in` and `tmmx_in` from regression anlaysis.
	* they are included in gdd

```{r}
# /*===== preparation =====*/
# /*---- reg_data ----*/
gene_vars <- c("nrdname", "year")
weather_var <- c("pr_in", "pet_in", "gdd_in")
soil_var <- c("sandtotal_r", "claytotal_r", "silttotal_r", "ksat_r", "awc_r", "slope_r")

reg_long_dt <- reg_data %>%
	.[,c(gene_vars, weather_var, soil_var), with=FALSE] %>%
	melt(id.vars = gene_vars)

 
# /*=================================================*/
#' # Summary statistics 
# /*=================================================*/
reg_data[,c("nrdname", weather_var, soil_var), with = FALSE] %>%
	datasummary(
		formula = as.formula( paste(paste0(c(weather_var, soil_var), collapse = "+"), "nrdname * (Mean + SD + MinMax)",  sep="~")),
		output = "gt", data = .
		)
```

+ unit:
	* `pr_in`(inch)
	* `tmmn_in` and `tmmx_in`(celsius)
	* `pet_in`(inch)
	* `sandtotal_r`, `claytotal_r`, `silttotal_r` (%)
	* `ksat_r` (um/s)
	* `awc_r` (cm/cm)
	* `slope_r` (%)

+ 

<br>

```{r, fig.width = 7.5, fig.height = 7.5, out.width = "750px", out.height = "750px"}
# /*=================================================*/
#' # Quick Visualization
# /*=================================================*/
# /*===== Comparison: Treatment and Control groups =====*/
# /*---- weather  ----*/
ggplot(reg_long_dt[variable %in% weather_var,]) +
	geom_boxplot(aes(x=factor(year), y=value, fill=nrdname)) +
	facet_grid(variable ~ . , scale='free') +
	theme(
		legend.position = "bottom",
		title = element_text(face = "bold"),
		strip.text.y = element_text(face = "bold")
		) +
	ggtitle("Annual Variations in Climate Variables")
```

```{r}
# /*---- soil  ----*/
ggplot(reg_long_dt[variable %in% soil_var,]) +
	geom_histogram(aes(x=value))+
	facet_grid(nrdname ~ variable ,scale='free') +
	ggtitle("Histogram by Soil Variables") +
	theme(
		title = element_text(face = "bold"),
		strip.text.y = element_text(face = "bold"),
		strip.text.x = element_text(face = "bold")
		)
```

+ Similar variations in the weather and soil variables between treatment and control group. 

+ 

<br>

<br>

## Scatterplot matrix
 
```{r}
# /*=================================================*/
#' # Correlation Matrix
# /*=================================================*/
# /*---- weather ----*/
GGally::ggpairs(
	data = reg_data[,..weather_var],
	# lower = list(continuous = wrap("points", size=0.5)),
	diag= list(continuous="barDiag"),
	title = "Scatterplot matrix of the Weather Data"
	)
```


<br>

<br>

```{r}
# # /*---- soil ----*/
ggpairs(
	data = reg_data[,..soil_var],
	# lower = list(continuous = wrap("points", color= "blue", size=0.5)),
	diag=list(continuous="barDiag"),
	title = "Scatterplot matrix of the Soil Data"
	)
```
